move realtime hook to its own file + dir
